refactor: typing public interfaces with order by
This commit updates the public `.pyi` files to ensure we use the new
order by spec
