+++ dbtsl/client/asyncio.pyi
@@ -1,10 +1,10 @@
 from contextlib import AbstractAsyncContextManager
-from typing import AsyncIterator, List, Optional
+from typing import AsyncIterator, List, Optional, Union
 
 import pyarrow as pa
 from typing_extensions import Self, Unpack, overload
 
-from dbtsl.api.adbc.protocol import QueryParameters
+from dbtsl.api.shared.query_params import OrderByGroupBy, OrderByMetric, QueryParameters
 from dbtsl.models import Dimension, Entity, Measure, Metric, SavedQuery
 
 class AsyncSemanticLayerClient:
@@ -20,7 +20,7 @@ class AsyncSemanticLayerClient:
         metrics: List[str],
         group_by: Optional[List[str]] = None,
         limit: Optional[int] = None,
-        order_by: Optional[List[str]] = None,
+        order_by: Optional[List[Union[str, OrderByGroupBy, OrderByMetric]]] = None,
         where: Optional[List[str]] = None,
         read_cache: bool = True,
     ) -> str: ...
@@ -29,7 +29,7 @@ class AsyncSemanticLayerClient:
         self,
         saved_query: str,
         limit: Optional[int] = None,
-        order_by: Optional[List[str]] = None,
+        order_by: Optional[List[Union[OrderByGroupBy, OrderByMetric]]] = None,
         where: Optional[List[str]] = None,
         read_cache: bool = True,
     ) -> str: ...
@@ -43,7 +43,7 @@ class AsyncSemanticLayerClient:
         metrics: List[str],
         group_by: Optional[List[str]] = None,
         limit: Optional[int] = None,
-        order_by: Optional[List[str]] = None,
+        order_by: Optional[List[Union[str, OrderByGroupBy, OrderByMetric]]] = None,
         where: Optional[List[str]] = None,
         read_cache: bool = True,
     ) -> "pa.Table": ...
@@ -52,7 +52,7 @@ class AsyncSemanticLayerClient:
         self,
         saved_query: str,
         limit: Optional[int] = None,
-        order_by: Optional[List[str]] = None,
+        order_by: Optional[List[Union[OrderByGroupBy, OrderByMetric]]] = None,
         where: Optional[List[str]] = None,
         read_cache: bool = True,
     ) -> "pa.Table": ...
