+++ tests/integration/test_sl_client.py
@@ -3,7 +3,7 @@ from typing import AsyncIterator, Iterator, Union
 import pytest
 from pytest_subtests import SubTests
 
-from dbtsl import OrderByDimension
+from dbtsl import OrderByGroupBy
 from dbtsl.client.asyncio import AsyncSemanticLayerClient
 from dbtsl.client.base import ADBC, GRAPHQL
 from dbtsl.client.sync import SyncSemanticLayerClient
@@ -106,7 +106,7 @@ async def test_client_query_saved_query(api: str, client: BothClients) -> None:
     table = await maybe_await(
         client.query(
             saved_query="order_metrics",
-            order_by=["metric_time"],
+            order_by=[OrderByGroupBy(name="metric_time", grain=None)],
             where=["1=1"],
             limit=1,
             read_cache=True,
@@ -140,7 +140,7 @@ async def test_client_compile_sql_saved_query(client: BothClients) -> None:
     sql = await maybe_await(
         client.compile_sql(
             saved_query="order_metrics",
-            order_by=["metric_time"],
+            order_by=[OrderByGroupBy(name="metric_time", grain=None)],
             where=["1=1"],
             limit=1,
             read_cache=True,
